skills/qa/requirements.txt
pyyaml>=6.0
httpx>=0.27
jsonschema>=4.21
